+++ NEWS.md
@@ -11,7 +11,7 @@
 
 - Ace 1.43.5
 - Copilot Language Server 1.459.0
-- Electron 39.8.7
+- Electron 41.3.0
 - Node.js 22.22.2 (copilot, Posit AI)
 - Quarto 1.9.36
 - xterm.js 6.0.0